requirements.txt
efficientunet_pytorch==0.0.6
matplotlib==3.5.2
numpy==1.21.2
opencv_python==4.6.0.66
opencv_python_headless==4.6.0.66
scikit_learn==1.0.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
